create, delete, update, get update

+++ src/main/java/com/example/Sorokin_Event/dto/EventsDto.java
@@ -28,7 +28,7 @@ public record EventsDto(
             int cost,
             @NotNull
             int duration,
-            int locationId,
+            Long locationId,
             EventStatus status
 ) {
 }
